# Docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Printing API \u2014 v1",
        "version": "1.0.0",
        "description": "The **Printing** API. This page is generated from the source code on every request \u2014\ncontrollers, FormRequests and API Resources are the only definition of the contract,\nso what you read here is always what the code does right now.\n\n**Every response uses the same envelope:**\n\n```json\n{ \"status\": true, \"message\": \"\u062a\u0645 \u0628\u0646\u062c\u0627\u062d\", \"data\": {} }\n```\n\nValidation failures (422) add an `errors` object keyed by field name. Messages are Arabic.\n\n**To try protected endpoints:** call `POST /auth/login`, copy the `data.token` from the\nresponse, then click **Authorize** at the top of this page and paste it as the Bearer token.\nSeeded local account: `admin@printing.ly` / password `password`."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/v1/products/{product}/images/{image}": {
            "put": {
                "operationId": "products.images.update",
                "description": "Changes the caption, the ordering, or promotes this image to be the product's primary.\nThe file itself cannot be swapped \u2014 upload a new image and delete the old one, so a URL\nalready handed out never starts pointing at different content.",
                "summary": "Update an image",
                "tags": [
                    "ProductImage"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "\u062a\u0645 \u062a\u062d\u062f\u064a\u062b \u0627\u0644\u0635\u0648\u0631\u0629 \u0628\u0646\u062c\u0627\u062d"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductImageResource"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "products.images.destroy",
                "description": "Removes the record and the stored file. If it was the primary image, the next one takes\nits place so the product is never left without a thumbnail.",
                "summary": "Delete an image",
                "tags": [
                    "ProductImage"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "\u062a\u0645 \u062d\u0630\u0641 \u0627\u0644\u0635\u0648\u0631\u0629 \u0628\u0646\u062c\u0627\u062d"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "ProductImageResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "url": {
                        "type": "string",
                        "description": "Generated per request from the disk the file lives on, never stored \u2014 so it stays\ncorrect after a move to S3, and can be a signed link for a private bucket."
                    },
                    "alt_text": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "is_primary": {
                        "type": "boolean"
                    },
                    "sort_order": {
                        "type": "integer"
                    },
                    "mime_type": {
                        "type": "string"
                    },
                    "size_bytes": {
                        "type": "integer"
                    },
                    "width_px": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "height_px": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "url",
                    "alt_text",
                    "is_primary",
                    "sort_order",
                    "mime_type",
                    "size_bytes",
                    "width_px",
                    "height_px",
                    "created_at"
                ],
                "title": "ProductImageResource"
            },
...
            "UpdateProductImageRequest": {
                "type": "object",
                "description": "Edits an image's metadata. The file itself is never replaced \u2014 upload a new one and delete\nthe old, so a cached URL can never start pointing at different content.",
                "properties": {
                    "alt_text": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "sort_order": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 65535
                    },
                    "is_primary": {
                        "description": "Only promoting is meaningful: a product needs one primary image, so demoting the\ncurrent one without naming a replacement is refused.",
                        "enum": [
                            "yes",
                            "on",
                            "1",
                            1,
                            "true",
                            true
                        ]
                    }
                },
                "title": "UpdateProductImageRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}